Remove erroneous cast to (guint) on comparing -1 to status->parcent (which
2005-07-14  Ryan Lortie  <desrt@desrt.ca>

        * power-management.c (power_management_getinfo): Remove erroneous
          cast to (guint) on comparing -1 to status->parcent (which is
          actually signed).  Patch from Benoît Dejean <TazForEver@dlfp.org>.

+++ battstat/power-management.c
@@ -398,7 +398,7 @@ power_management_getinfo( BatteryStatus *status )
 
   retval = apm_readinfo( status );
 
-  if(status->percent == (guint)-1) {
+  if(status->percent == -1) {
     status->percent = 0;
     status->present = FALSE;
   }
